requirements.txt
Flask==1.1.1
config==0.4.2
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
psycopg2-binary==2.8.4
numpy==1.15.1
pandas==0.23.4 
scipy==1.1.0
matplotlib==2.2.2
slackclient==1.3.2
jira==2.0.0
seaborn==0.9.0
